Started implementing shop interface:  Buy Basic Items almost running.

+++ src/struct.h
@@ -347,7 +347,7 @@ typedef struct
   float RechargeTimeModifier; // the factor to be multiplied to the recharge time (times 100 for % notation)
   float RechargeTime; // the actual minimum time to wait for the influencer between shots
   int exp_level; // which 'experience level' is the influencer currenly at?
-  int ExpRequired; // how much experience required for the next level?
+  long ExpRequired; // how much experience required for the next level?
   long Gold;
   char character_name[25];
   gps Position_History_Ring_Buffer[ MAX_INFLU_POSITION_HISTORY ];
